9-SiamCAR/SiamCAR-V2/requirement.txt
numpy
pytorch==1.2.0
opencv-python
pyyaml
yacs
tqdm
colorama
matplotlib
cython
tensorboardX
